pyproject.toml
[project]
name = "mypythonprojects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gdal>=3.11.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyqt6>=6.9.1",
    "pyside6>=6.9.1",
    "rasterio>=1.4.3",
    "segmentation-models-pytorch>=0.5.0",
    "torch>=2.7.1",
]

[tool.uv.sources]
gdal = { path = "gdal-3.11.1-cp313-cp313-win_amd64.whl" } # windows only
